statement: split create_table.go helpers, rename, and add docs
Completes the reorganization:
  - rename parse_create_table.go -> create_table.go (and its test) so the
    file holding the CreateTable receivers is named for the type
  - extract the non-receiver helpers into parse_helpers.go (AST / type-string
    helpers), accessors.go (collection Has*/ByName/GetName), and
    secondary_indexes.go (RemoveSecondaryIndexes / GetMissingSecondaryIndexes)
  - move the secondary-index tests into secondary_indexes_test.go
  - document the normalization framework in pkg/statement/README.md and AGENTS.md

Pure move + docs; no behavior change. With the file layout and filenames now
final, the docs describe them accurately.

Stacked on #1056.
