+++ pkg/statement/create_table.go
@@ -8,12 +8,10 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/block/spirit/pkg/dbconn/sqlescape"
 	"github.com/pingcap/tidb/pkg/parser"
 	"github.com/pingcap/tidb/pkg/parser/ast"
 	"github.com/pingcap/tidb/pkg/parser/format"
 	"github.com/pingcap/tidb/pkg/parser/mysql"
-	driver "github.com/pingcap/tidb/pkg/parser/test_driver"
 	"github.com/pingcap/tidb/pkg/parser/types"
 )
 
@@ -334,26 +332,6 @@ func (ct *CreateTable) GetTableOptions() map[string]any {
 
 func (ct *CreateTable) GetPartition() *PartitionOptions {
 	return ct.Partition
-}
-
-func (indexes Indexes) HasInvisible() bool {
-	for _, idx := range indexes {
-		if idx.Invisible != nil && *idx.Invisible {
-			return true
-		}
-	}
-
-	return false
-}
-
-func (constraints Constraints) HasForeignKeys() bool {
-	for _, c := range constraints {
-		if c.Type == "FOREIGN KEY" {
-			return true
-		}
-	}
-
-	return false
 }
 
 // parseToStruct converts the AST into a structured CreateTable
@@ -1129,79 +1107,6 @@ func (ct *CreateTable) parseSubPartitionDefinition(sub *ast.SubPartitionDefiniti
 	return subDef
 }
 
-// stringLiteralValue returns the true, fully-unescaped value of a quoted
-// string-literal AST node (for example the value behind a single-quoted
-// DEFAULT, or a COMMENT, whose contents include escaped quote or backslash
-// characters) along with true. For any other expression kind it returns an
-// empty string and false.
-//
-// This is the load-bearing fix for the string round-trip bugs: the TiDB
-// parser's Restore re-emits a string literal in its re-escaped, still-quoted
-// form rather than as the raw value, so the previous approach of
-// Restore-then-strip-outer-quotes left the inner escaping in place. Reading
-// the literal's value directly off the AST yields the raw bytes, which we
-// then escape exactly once at emission time via sqlescape (backslash
-// escaping, which MySQL accepts in its default sql_mode). Note MySQL renders
-// a literal quote in SHOW CREATE TABLE as a doubled quote, which the parser
-// also accepts; the doubled and backslash forms are equivalent in the
-// default sql_mode.
-func stringLiteralValue(expr ast.ExprNode) (string, bool) {
-	if v, ok := expr.(*driver.ValueExpr); ok && v.Kind() == driver.KindString {
-		return v.GetString(), true
-	}
-	return "", false
-}
-
-// isExpressionDefault returns true when the default value expression should be
-// wrapped in parentheses in the generated DDL. MySQL requires expression defaults
-// (as opposed to literal defaults) to be enclosed in parens, e.g. DEFAULT (json_object()).
-// This mirrors the logic in the TiDB parser's ColumnOption.Restore for ColumnOptionDefaultValue:
-// non-CURRENT_TIMESTAMP function calls and column name expressions get outer parentheses.
-func isExpressionDefault(expr ast.ExprNode) bool {
-	if expr == nil {
-		return false
-	}
-	switch e := expr.(type) {
-	case *ast.FuncCallExpr:
-		// CURRENT_TIMESTAMP (and aliases NOW, LOCALTIME, etc.) are literal-style defaults
-		// that don't need parens. Everything else is an expression default.
-		switch e.FnName.L {
-		case "current_timestamp", "now", "localtime", "localtimestamp", "utc_timestamp":
-			return false
-		}
-		return true
-	case *ast.ColumnNameExpr:
-		return true
-	default:
-		return false
-	}
-}
-
-// restoreExpressionText restores an expression AST node to its SQL text,
-// stripping redundant outer parentheses. MySQL's SHOW CREATE TABLE wraps
-// generated-column and CHECK expressions in an extra set of parentheses
-// (e.g. GENERATED ALWAYS AS ((`a` + 1))); stripping them ensures a
-// user-written `AS (a + 1)` compares equal to the canonical form.
-// Unlike parseExpression, the result is NOT lowercased and string literals
-// keep their quotes — these expressions may contain case-sensitive literals.
-func restoreExpressionText(expr ast.ExprNode) (string, bool) {
-	for {
-		paren, ok := expr.(*ast.ParenthesesExpr)
-		if !ok {
-			break
-		}
-		expr = paren.Expr
-	}
-
-	var sb strings.Builder
-	rCtx := format.NewRestoreCtx(format.DefaultRestoreFlags|format.RestoreStringWithoutCharset, &sb)
-	if err := expr.Restore(rCtx); err != nil {
-		return "", false
-	}
-
-	return sb.String(), true
-}
-
 // parseExpression converts an expression to a string representation
 func (ct *CreateTable) parseExpression(expr ast.ExprNode) any {
 	if expr == nil {
@@ -1255,277 +1160,4 @@ func (ct *CreateTable) parseExpression(expr ast.ExprNode) any {
 		}
 		return str
 	}
-}
-
-// extractLengthFromTypeString extracts length from type string like "varchar(100)"
-func extractLengthFromTypeString(typeStr string) int {
-	// Simple regex-like parsing for common cases
-	if strings.Contains(typeStr, "(") && strings.Contains(typeStr, ")") {
-		start := strings.Index(typeStr, "(")
-
-		end := strings.Index(typeStr, ")")
-		if start < end && start != -1 && end != -1 {
-			lengthStr := typeStr[start+1 : end]
-			// Handle cases like "decimal(10,2)" - take the first number
-			if commaIdx := strings.Index(lengthStr, ","); commaIdx != -1 {
-				lengthStr = lengthStr[:commaIdx]
-			}
-
-			var length int
-			if n, err := fmt.Sscanf(lengthStr, "%d", &length); n == 1 && err == nil {
-				return length
-			}
-		}
-	}
-
-	return 0
-}
-
-// extractPrecisionScaleFromTypeString extracts precision and scale from type string like "decimal(10,2)"
-func extractPrecisionScaleFromTypeString(typeStr string) (int, int) {
-	if strings.Contains(typeStr, "(") && strings.Contains(typeStr, ")") {
-		start := strings.Index(typeStr, "(")
-
-		end := strings.Index(typeStr, ")")
-		if start < end && start != -1 && end != -1 {
-			paramStr := typeStr[start+1 : end]
-			if precisionStr, scaleStr, found := strings.Cut(paramStr, ","); found {
-				precisionStr = strings.TrimSpace(precisionStr)
-				scaleStr = strings.TrimSpace(scaleStr)
-
-				var precision, scale int
-				if n, err := fmt.Sscanf(precisionStr, "%d", &precision); n == 1 && err == nil {
-					if n, err := fmt.Sscanf(scaleStr, "%d", &scale); n == 1 && err == nil {
-						return precision, scale
-					}
-
-					return precision, 0
-				}
-			}
-		}
-	}
-
-	return 0, 0
-}
-
-// ByName is a generic function that finds an element by name in any slice of types with Name field
-// NOTE: This function assumes that names are unique within the slice! That will be true for
-// "canonical" CREATE TABLE statements as returned by SHOW CREATE TABLE, but may not be true for
-// arbitrary input.
-func ByName[T HasName](slice []T, name string) *T {
-	for _, item := range slice {
-		if item.GetName() == name {
-			return &item
-		}
-	}
-
-	return nil
-}
-
-func (i Index) GetName() string {
-	return i.Name
-}
-
-func (c Column) GetName() string {
-	return c.Name
-}
-
-func (c Constraint) GetName() string {
-	return c.Name
-}
-
-func (indexes Indexes) ByName(name string) *Index {
-	return ByName(indexes, name)
-}
-
-func (columns Columns) ByName(name string) *Column {
-	return ByName(columns, name)
-}
-
-func (constraints Constraints) ByName(name string) *Constraint {
-	return ByName(constraints, name)
-}
-
-// RemoveSecondaryIndexes takes a CREATE TABLE statement and returns a modified version
-// without secondary indexes (regular INDEX only). PRIMARY KEY, UNIQUE, and FULLTEXT
-// indexes are preserved.
-func RemoveSecondaryIndexes(createStmt string) (string, error) {
-	ct, err := ParseCreateTable(createStmt)
-	if err != nil {
-		return "", fmt.Errorf("failed to parse CREATE TABLE: %w", err)
-	}
-
-	// Filter out regular INDEX entries from the constraints
-	filteredConstraints := make([]*ast.Constraint, 0)
-	for _, constraint := range ct.Raw.Constraints {
-		// Keep everything except regular INDEX
-		switch constraint.Tp {
-		case ast.ConstraintKey, ast.ConstraintIndex:
-			// Skip regular indexes
-			continue
-		default:
-			// Keep PRIMARY KEY, UNIQUE, FULLTEXT, SPATIAL, etc.
-			filteredConstraints = append(filteredConstraints, constraint)
-		}
-	}
-
-	// Update the AST with filtered constraints
-	ct.Raw.Constraints = filteredConstraints
-
-	// Restore the modified AST back to SQL
-	var sb strings.Builder
-	rCtx := format.NewRestoreCtx(format.DefaultRestoreFlags, &sb)
-	if err := ct.Raw.Restore(rCtx); err != nil {
-		return "", fmt.Errorf("failed to restore CREATE TABLE: %w", err)
-	}
-
-	return sb.String(), nil
-}
-
-// GetMissingSecondaryIndexes compares two CREATE TABLE statements (source and target)
-// and returns an ALTER TABLE statement that adds any missing secondary indexes.
-// Returns an empty string if no indexes need to be added.
-// Considers UNIQUE, FULLTEXT, SPATIAL, and regular INDEX types. PRIMARY KEY is excluded as it's fundamental to table structure.
-func GetMissingSecondaryIndexes(sourceCreateTable, targetCreateTable, tableName string) (string, error) {
-	// Parse both CREATE TABLE statements
-	sourceCT, err := ParseCreateTable(sourceCreateTable)
-	if err != nil {
-		return "", fmt.Errorf("failed to parse source CREATE TABLE: %w", err)
-	}
-
-	targetCT, err := ParseCreateTable(targetCreateTable)
-	if err != nil {
-		return "", fmt.Errorf("failed to parse target CREATE TABLE: %w", err)
-	}
-
-	// Build a map of existing indexes on the target (all secondary indexes)
-	targetIndexes := make(map[string]bool)
-	for _, constraint := range targetCT.Raw.Constraints {
-		// Skip PRIMARY KEY, but include all other index types
-		if constraint.Tp == ast.ConstraintPrimaryKey {
-			continue
-		}
-		targetIndexes[constraint.Name] = true
-	}
-
-	// Find missing indexes from source
-	var missingIndexes []*ast.Constraint
-	for _, constraint := range sourceCT.Raw.Constraints {
-		// Consider all secondary indexes (UNIQUE, FULLTEXT, and regular INDEX)
-		// Skip only PRIMARY KEY as it's fundamental to table structure
-		if constraint.Tp == ast.ConstraintPrimaryKey {
-			continue
-		}
-		// Include: UNIQUE, FULLTEXT, SPATIAL, and regular INDEX
-		if constraint.Tp != ast.ConstraintKey &&
-			constraint.Tp != ast.ConstraintIndex &&
-			constraint.Tp != ast.ConstraintUniq &&
-			constraint.Tp != ast.ConstraintUniqKey &&
-			constraint.Tp != ast.ConstraintUniqIndex &&
-			constraint.Tp != ast.ConstraintFulltext &&
-			constraint.Tp != ast.ConstraintSpatial {
-			continue
-		}
-
-		// Check if this index exists on target
-		if !targetIndexes[constraint.Name] {
-			missingIndexes = append(missingIndexes, constraint)
-		}
-	}
-
-	// If no missing indexes, return empty string
-	if len(missingIndexes) == 0 {
-		return "", nil
-	}
-
-	// Build a single ALTER TABLE statement with all missing indexes
-	var alterClauses []string
-	for _, constraint := range missingIndexes {
-		var sb strings.Builder
-
-		// Add appropriate keyword based on constraint type
-		switch constraint.Tp {
-		case ast.ConstraintUniq, ast.ConstraintUniqKey, ast.ConstraintUniqIndex:
-			sb.WriteString("ADD UNIQUE INDEX")
-		case ast.ConstraintFulltext:
-			sb.WriteString("ADD FULLTEXT INDEX")
-		case ast.ConstraintSpatial:
-			sb.WriteString("ADD SPATIAL INDEX")
-		default: // ast.ConstraintKey, ast.ConstraintIndex
-			sb.WriteString("ADD INDEX")
-		}
-
-		// Add index name if present
-		if constraint.Name != "" {
-			fmt.Fprintf(&sb, " %s", sqlescape.EscapeIdentifier(constraint.Name))
-		}
-
-		// Add columns
-		sb.WriteString(" (")
-		for i, key := range constraint.Keys {
-			if i > 0 {
-				sb.WriteString(", ")
-			}
-			switch {
-			case key.Column != nil:
-				// Regular column reference
-				sb.WriteString(sqlescape.EscapeIdentifier(key.Column.Name.String()))
-				// Add length if specified
-				if key.Length > 0 {
-					fmt.Fprintf(&sb, "(%d)", key.Length)
-				}
-			case key.Expr != nil:
-				// Functional (expression) index part, e.g. KEY ((lower(b))).
-				// Column is nil in this case; MySQL requires the expression
-				// to be wrapped in its own parentheses.
-				var exprSb strings.Builder
-				rCtx := format.NewRestoreCtx(format.DefaultRestoreFlags|format.RestoreStringWithoutCharset, &exprSb)
-				if err := key.Expr.Restore(rCtx); err != nil {
-					return "", fmt.Errorf("failed to restore expression for index %q: %w", constraint.Name, err)
-				}
-				fmt.Fprintf(&sb, "(%s)", exprSb.String())
-			default:
-				return "", fmt.Errorf("index %q has a key part with neither a column nor an expression", constraint.Name)
-			}
-			// Descending key part (MySQL 8.0+). ASC is MySQL's canonical
-			// default and is never emitted explicitly.
-			if key.Desc {
-				sb.WriteString(" DESC")
-			}
-		}
-		sb.WriteString(")")
-
-		// Add index options if present
-		if constraint.Option != nil {
-			opt := constraint.Option
-
-			// Add USING clause
-			if opt.Tp != ast.IndexTypeInvalid && opt.Tp.String() != "" {
-				sb.WriteString(" USING " + opt.Tp.String())
-			}
-
-			// Add KEY_BLOCK_SIZE
-			if opt.KeyBlockSize > 0 {
-				fmt.Fprintf(&sb, " KEY_BLOCK_SIZE=%d", opt.KeyBlockSize)
-			}
-
-			// Add WITH PARSER (FULLTEXT indexes)
-			if opt.ParserName.L != "" {
-				fmt.Fprintf(&sb, " WITH PARSER %s", opt.ParserName.String())
-			}
-
-			// Add COMMENT
-			if opt.Comment != "" {
-				fmt.Fprintf(&sb, " COMMENT '%s'", sqlescape.EscapeString(opt.Comment))
-			}
-
-			// Add VISIBLE/INVISIBLE
-			if opt.Visibility == ast.IndexVisibilityInvisible {
-				sb.WriteString(" INVISIBLE")
-			}
-		}
-		alterClauses = append(alterClauses, sb.String())
-	}
-	// Combine all ADD INDEX clauses into a single ALTER TABLE statement
-	return fmt.Sprintf("ALTER TABLE %s %s", sqlescape.EscapeIdentifier(tableName), strings.Join(alterClauses, ", ")), nil
 }
